requirements.txt
pandas==2.2.3
azure-ai-documentintelligence==1.0.2
azure-ai-formrecognizer==3.3.3
azure-common==1.1.28
azure-core==1.34.0
streamlit==1.45.0